requirements.txt
# Core Framework
Django>=5.0.0,<5.1.0
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.0

# Image Processing
Pillow>=10.0.0

# Async Task Queue
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0

# WebSocket Support
channels>=4.0.0
channels-redis>=4.0.0
daphne>=4.0.0

# Google Cloud AI
google-cloud-aiplatform>=1.38.0
google-auth>=2.25.0
google-genai>=0.3.0

# API & Web
gunicorn>=21.0.0

# Development Tools (optional)
django-extensions>=3.2.0
ipython>=8.0.0
